AeroDocSense/requirements.txt
streamlit==1.45.1
python-dotenv==1.0.1
huggingface_hub==0.33.4
pymongo==4.6.0
numpy==2.0.1
pydantic_settings==2.10.1
fastapi==0.116.1
pyMuPDF==1.26.3
scikit-learn==1.5.1
dnspython==2.4.2